chore: verify WSL release readiness for validation workflow
Wire live drivers to real protocol stacks (rusty-modbus, reqwest JSON parse),
stop Haystack auto-fixture and BACnet live scan registry fallback, fix clippy
clean lib build, and add Linux runtime handoff doc for PR #385.

+++ edge/src/drivers/haystack/parity.rs
@@ -40,7 +40,7 @@ pub fn compare_reading(input: &ParityInput, tol: &ParityTolerance) -> Value {
     } else {
         tol.temperature_f
     };
-    let pass = abs_delta <= tolerance && ts_delta <= tol.timestamp_skew_seconds as i64;
+    let pass = abs_delta <= tolerance && ts_delta <= tol.timestamp_skew_seconds;
     json!({
         "equipment_id": input.equipment_id,
         "role": input.role,
